split visitors into new file
